Addition of several libraries to satisfy the dependencies of GeoTools

+++ ummisco.gama.java2d/src/ummisco/gama/java2d/Java2DDisplaySurface.java
@@ -502,7 +502,7 @@ public class Java2DDisplaySurface extends JPanel implements IDisplaySurface {
 
 	@Override
 	public double getDisplayWidth() {
-		return (double) viewPort.width;
+		return viewPort.width;
 	}
 
 	protected void setDisplayWidth(final int displayWidth) {
@@ -516,7 +516,7 @@ public class Java2DDisplaySurface extends JPanel implements IDisplaySurface {
 
 	@Override
 	public double getDisplayHeight() {
-		return (double) viewPort.height;
+		return viewPort.height;
 	}
 
 	protected void setDisplayHeight(final int displayHeight) {
